App2
Here a random number defines the String color using a switch statement. Once the color value is assigned, a scanner is implemented to request the user to input the color that was randomly selected which is nested in a do-while loop in the case that the user enters the wrong input. Once the correct color is input, the scanner closes and a message is printed signalling success.